+++ matrix-sdk/src/main/java/org/matrix/androidsdk/RestClient.java
@@ -62,7 +62,13 @@ public class RestClient<T> {
      * Prefix used in path of identity server API requests.
      */
     public static final String URI_IDENTITY_PATH = "_matrix/identity/api/v1";
+    public static final String URI_IDENTITY_PATH_V2 = "_matrix/identity/v2/";
     public static final String URI_API_PREFIX_IDENTITY = URI_IDENTITY_PATH + "/";
+
+    /**
+     * Prefix used for integration manager
+     */
+    public static final String URI_INTEGRATION_MANAGER_PATH = "_matrix/integrations/v1";
 
     /**
      * List the servers which should be used to define the base url.